# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Captora Photography CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all client payments
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Add a payment installment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /freelancers:
    get:
      operationId: listFreelancers
      tags: [freelancers]
      summary: List all freelancers
      responses:
        "200":
          description: List of freelancers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Freelancer"
    post:
      operationId: createFreelancer
      tags: [freelancers]
      summary: Add a freelancer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreelancerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Freelancer"
  /freelancers/{id}:
    get:
      operationId: getFreelancer
      tags: [freelancers]
      summary: Get freelancer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Freelancer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerDetail"
    patch:
      operationId: updateFreelancer
      tags: [freelancers]
      summary: Update a freelancer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreelancerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Freelancer"
    delete:
      operationId: deleteFreelancer
      tags: [freelancers]
      summary: Delete a freelancer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /freelancers/{id}/payments:
    get:
      operationId: listFreelancerPayments
      tags: [freelancers]
      summary: Get payment history for a freelancer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Freelancer payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FreelancerPayment"
    post:
      operationId: addFreelancerPayment
      tags: [freelancers]
      summary: Record a freelancer payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreelancerPaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerPayment"

  # STAFF
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List all expenses
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Add an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /expenses/summary:
    get:
      operationId: getExpenseSummary
      tags: [expenses]
      summary: Get monthly/yearly expense summary
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Expense summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseSummary"

  # P&L
...
components:
  schemas:
...
    Payment:
      type: object
      required: [id, clientId, amount, paymentDate, installmentType, mode, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        amount:
          type: number
        paymentDate:
          type: string
        installmentType:
          type: string
        mode:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [clientId, amount, paymentDate, installmentType, mode]
      properties:
        clientId:
          type: integer
        amount:
          type: number
        paymentDate:
          type: string
        installmentType:
          type: string
        mode:
          type: string
        notes:
          type: string
    PaymentUpdate:
      type: object
      properties:
        amount:
          type: number
        paymentDate:
          type: string
        installmentType:
          type: string
        mode:
          type: string
        notes:
          type: string
...
    Freelancer:
      type: object
      required: [id, name, role, phone, perShootRate, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        bankDetails:
          type: ["string", "null"]
        perShootRate:
          type: number
        totalEarned:
          type: number
        totalPaid:
          type: number
        totalDue:
          type: number
        createdAt:
          type: string
    FreelancerDetail:
      allOf:
        - $ref: "#/components/schemas/Freelancer"
        - type: object
          properties:
            payments:
              type: array
              items:
                $ref: "#/components/schemas/FreelancerPayment"
    FreelancerInput:
      type: object
      required: [name, role, phone, perShootRate]
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        bankDetails:
          type: string
        perShootRate:
          type: number
    FreelancerUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        bankDetails:
          type: string
        perShootRate:
          type: number
    FreelancerPayment:
      type: object
      required: [id, freelancerId, amount, paymentDate, status, createdAt]
      properties:
        id:
          type: integer
        freelancerId:
          type: integer
        shootId:
          type: ["integer", "null"]
        amount:
          type: number
        paymentDate:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    FreelancerPaymentInput:
      type: object
      required: [amount, paymentDate, status]
      properties:
        shootId:
          type: integer
        amount:
          type: number
        paymentDate:
          type: string
        status:
          type: string
        notes:
          type: string
...
    Expense:
      type: object
      required: [id, category, amount, expenseDate, paidBy, createdAt]
      properties:
        id:
          type: integer
        category:
          type: string
        amount:
          type: number
        expenseDate:
          type: string
        description:
          type: ["string", "null"]
        paidBy:
          type: string
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [category, amount, expenseDate, paidBy]
      properties:
        category:
          type: string
        amount:
          type: number
        expenseDate:
          type: string
        description:
          type: string
        paidBy:
          type: string
    ExpenseUpdate:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        expenseDate:
          type: string
        description:
          type: string
        paidBy:
          type: string
    ExpenseSummary:
      type: object
      required: [totalExpenses, byCategory]
      properties:
        totalExpenses:
          type: number
        byCategory:
          type: array
          items:
            type: object
            required: [category, total]
            properties:
              category:
                type: string
              total:
                type: number
...