# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Captora Photography CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  category: string;
  amount: number;
  expenseDate: string;
  /** @nullable */
  description?: string | null;
  paidBy: string;
  createdAt: string;
}