# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Captora Photography CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseUpdate {
  category?: string;
  amount?: number;
  expenseDate?: string;
  description?: string;
  paidBy?: string;
}